Some more refactoring.
