+++ src/main/scala/edu/ie3/simona/agent/grid/ReceivedValuesStore.scala
@@ -7,7 +7,6 @@
 package edu.ie3.simona.agent.grid
 
 import edu.ie3.datamodel.graph.SubGridGate
-import edu.ie3.simona.agent.grid.GridAgentMessages.SlackVoltageResponse
 import edu.ie3.simona.agent.grid.GridAgentMessages.SlackVoltageResponse.ExchangeVoltage
 import edu.ie3.simona.agent.grid.ReceivedValuesStore.{
   NodeToReceivedPower,
@@ -48,53 +47,10 @@ final case class ReceivedValuesStore private (
 
 object ReceivedValuesStore {
 
-  sealed trait ReceivedValues
-
   type NodeToReceivedPower =
     Map[UUID, Map[ActorRef[_], Option[PowerResponseMessage]]]
   type NodeToReceivedSlackVoltage =
     Map[UUID, Option[ExchangeVoltage]]
-
-  private type ParticipantPowerRequestResponse =
-    (
-        ActorRef[_],
-        PowerResponseMessage,
-    ) // necessary, because participants are still classic actors
-  private type GridPowerRequestResponse =
-    (ActorRef[GridAgent.Request], PowerResponseMessage)
-  private type ActorSlackVoltageRequestResponse =
-    (ActorRef[GridAgent.Request], SlackVoltageResponse)
-
-  sealed trait ReceivedPowerValues extends ReceivedValues {
-    def values: Vector[(ActorRef[_], PowerResponseMessage)]
-  }
-
-  /** Wrapper for received asset power values (p, q)
-    *
-    * @param values
-    *   the asset power values and their senders
-    */
-  final case class ReceivedAssetPowerValues(
-      values: Vector[ParticipantPowerRequestResponse]
-  ) extends ReceivedPowerValues
-
-  /** Wrapper for received grid power values (p, q)
-    *
-    * @param values
-    *   the grid power values and their senders
-    */
-  final case class ReceivedGridPowerValues(
-      values: Vector[GridPowerRequestResponse]
-  ) extends ReceivedPowerValues
-
-  /** Wrapper for received slack voltage values (v)
-    *
-    * @param values
-    *   the slack voltage values and their senders
-    */
-  final case class ReceivedSlackVoltageValues(
-      values: Vector[ActorSlackVoltageRequestResponse]
-  ) extends ReceivedValues
 
   /** Get an empty, ready to be used instance of [[ReceivedValuesStore]]
     * containing an `empty` mapping of [[NodeToReceivedPower]] and
